microservices/linkedin_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
linkedin-api==2.3.1
pyotp==2.9.0
selenium==4.15.2
webdriver-manager==4.0.1 